# swagger/swagger.yaml
swagger: "2.0"
info:
  description: "External endpoints for erebor domain"
  version: "1.0.0"
  title: "Swagger Erebor"
host: "localhost:8080"
basePath: "/"
tags:
  - name: "status"
    description: "API status"
  - name: "greeting"
    description: "Greeting messages"
  - name: "account"
    description: "Bank accounts"
schemes:
  - "http"
paths:
  /status:
    get:
      tags:
        - "status"
      summary: "Current API status"
      description: ""
      operationId: "getStatus"
      produces:
        - "application/json"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/BaseResponse"
          examples:
            application/json:
              {
                "dateTime": "2020-09-30T13:05:28.128",
                "message": "UP!",
              }
  /greeting:
    post:
      tags:
        - "greeting"
      summary: "Says hello"
      description: "Say hello to a specific user"
      operationId: "createUser"
      parameters:
        - in: "body"
          name: "User"
          description: "Who receives the greeting"
          required: true
          schema:
            $ref: "#/definitions/User"
      produces:
        - "application/json"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/BaseResponse"
          examples:
            application/json:
              {
                "dateTime": "2020-09-30T13:05:20.708",
                "message": "Hello John Doe"
              }
  /{country}/account:
    get:
      tags:
        - "account"
      summary: "Find an account"
      description: ""
      operationId: "findAccount"
      parameters:
        - in: path
          name: country
          type: string
          required: true
          default: "USA"
          description: "Bank account's country code"
        - in: query
          name: no
          type: integer
          required: true
          default: 80001
          description: "Bank account's number"
      produces:
        - "application/json"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/AccountResponse"
          examples:
            application/json:
              {
                "dateTime": "2020-09-30T13:23:12.885",
                "account": {
                  "countryCode": "USA",
                  "no": 80001,
                  "balance": 5000000.0
                }
              }
definitions:
  User:
    type: object
    properties:
      name:
        type: string
        example: "John Doe"
  ResponseDto:
    type: object
    properties:
      dateTime:
        type: string
        example: "2020-09-30T13:29:00.001"
      message:
        type: string
        example: "UP!"

  AccountResponseDto:
    type: object
    properties:
      dateTime:
        type: string
        example: "2020-09-30T13:29:00.001"
      account:
        type: object
        properties:
          countryCode:
            type: string
            example: "USA"
          no:
            type: integer
            example: 80001
          balance:
            type: number
            example: 5000000.0